Add EXE006: Shebang is present but it hard-codes the path to python
The Python tutorial suggests using `#!/usr/bin/env python` for
shebangs.

This way is better than plain `#!/usr/bin/python` because it allows
the user running the script to control which Python executable to
interpret the script with.

This is especially true when working with virtualenvs, custom install
prefixes, and the like.